FA kernel tl.dot dtype 修复 (fp16 对齐引入的采集失败根因) + 工业级基准缺省自动跑 (AUTO_RUN_IND_BENCH)
- input/flash_attention/kernel_op.py: p=exp 结果(fp32) 与 vv(fp16) 的 tl.dot 输入 dtype 不匹配
  → kernel 编译失败 → op_summary 无目标 kernel → '采集不到 kernel 名'. 修复: P 转 fp16 再 dot
  (FA-2 标准, acc 累加保持 fp32), l_i 归一化分母求和转回 fp32 防精度丢失
- analyzers/sweep_blocks.py: FA sweep setup 输出 o 对齐 fp32 (与 kernel store dtype 一致)
- agents/scheduler.py: 新增 _load_industrial_best (mode 内 median / mode 间取最小 / 仅真执行,
  actual_mode==mode 回退不顶替) + _auto_run_industrial (缺哪个 mode 跑哪个 bench_industrial.py,
  env AUTO_RUN_IND_BENCH=0 关闭) → 缺工业级基准不再只警告, 自动生成后注入 vs_industrial_ratio
- README: ④ 工业级基准步骤说明同步自动跑逻辑
- 验证: FA dtype 静态断言 + _load_industrial_best 三场景 (回退不顶替/全缺/部分缺) + 回归 32/32 + flow/edge/字段链路全过

+++ triton_agent_optimizer/input/flash_attention/kernel_op.py
@@ -80,9 +80,13 @@ def flash_attn_mha_kernel(q_ptr, k_ptr, v_ptr, o_ptr,
         s = tl.where(causal, s, float("-inf"))
 
         m_curr = tl.maximum(tl.max(s, axis=1), m_i)   # running max [BLOCK_M]
-        p = tl.exp(s - m_curr[:, None])               # [BLOCK_M, BLOCK_N]
+        # ★P 转 fp16: tl.dot 两个输入必须同 dtype — p 是 exp 结果(fp32), vv 是 fp16 输入,
+        #   dot(fp32, fp16) 会编译失败 → kernel 起不来 → msprof 采不到 kernel 名.
+        #   FA 标准做法: P 矩阵用 fp16 (精度损失可忽略), acc 累加保持 fp32.
+        p = tl.exp(s - m_curr[:, None]).to(tl.float16)  # [BLOCK_M, BLOCK_N]
         alpha = tl.exp(m_i - m_curr)                  # 旧块衰减 [BLOCK_M]
-        l_i = alpha * l_i + tl.sum(p, axis=1)         # running sum [BLOCK_M]
+        # ★l_i 求和用 fp32 (p 已是 fp16, 直接 sum 会 fp16 累加掉精度; FA 标准: 归一化分母保持 fp32)
+        l_i = alpha * l_i + tl.sum(p.to(tl.float32), axis=1)  # running sum [BLOCK_M]
 
         # V[h, n, d]: h*(seq*dim) + n*dim + d  → 行跨度=1 (连续✓)
         v_ptrs = v_ptr + head * (seq * dim) + offs_n[:, None] * dim + offs_k[None, :]
